% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{AmericanSoccerAnalysis}
\alias{AmericanSoccerAnalysis}
\title{Class representing an active session connected to the American Soccer Analysis public API.}
\description{
Class representing an active session connected to the American Soccer Analysis public API.

Class representing an active session connected to the American Soccer Analysis public API.
}
\details{
Does not take any arguments to initialize.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{API_VERSION}}{Latest API version.}

\item{\code{LEAGUES}}{List of stylized league names.}

\item{\code{BASE_URL}}{API base URL.}

\item{\code{players}}{Data frame containing players from all leagues.}

\item{\code{teams}}{Data frame containing teams from all leagues.}

\item{\code{stadia}}{Data frame containing stadia from all leagues.}

\item{\code{managers}}{Data frame containing managers from all leagues.}

\item{\code{referees}}{Data frame containing referees from all leagues.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AmericanSoccerAnalysis$new()}}
\item \href{#method-get_players}{\code{AmericanSoccerAnalysis$get_players()}}
\item \href{#method-get_teams}{\code{AmericanSoccerAnalysis$get_teams()}}
\item \href{#method-get_stadia}{\code{AmericanSoccerAnalysis$get_stadia()}}
\item \href{#method-get_managers}{\code{AmericanSoccerAnalysis$get_managers()}}
\item \href{#method-get_referees}{\code{AmericanSoccerAnalysis$get_referees()}}
\item \href{#method-get_games}{\code{AmericanSoccerAnalysis$get_games()}}
\item \href{#method-clone}{\code{AmericanSoccerAnalysis$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{AmericanSoccerAnalysis} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AmericanSoccerAnalysis$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{AmericanSoccerAnalysis} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_players"></a>}}
\if{latex}{\out{\hypertarget{method-get_players}{}}}
\subsection{Method \code{get_players()}}{
Retrieves a data frame containing player names, IDs, and other metadata.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AmericanSoccerAnalysis$get_players(leagues, ids, names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{leagues}}{Leagues on which to filter. Accepts a character vector of length >= 1.}

\item{\code{ids}}{Player IDs on which to filter. Accepts a character vector of length >= 1.}

\item{\code{names}}{Player names on which to filter. Partial matches are accepted. Accepts a character vector of length >= 1.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_teams"></a>}}
\if{latex}{\out{\hypertarget{method-get_teams}{}}}
\subsection{Method \code{get_teams()}}{
Retrieves a data frame containing team names, abbreviations, and IDs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AmericanSoccerAnalysis$get_teams(leagues, ids, names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{leagues}}{Leagues on which to filter. Accepts a character vector of length >= 1.}

\item{\code{ids}}{Team IDs on which to filter. Accepts a character vector of length >= 1.}

\item{\code{names}}{Team names on which to filter. Partial matches and abbreviations are accepted. Accepts a character vector of length >= 1.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_stadia"></a>}}
\if{latex}{\out{\hypertarget{method-get_stadia}{}}}
\subsection{Method \code{get_stadia()}}{
Retrieves a data frame containing stadium names, IDs, and other metadata.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AmericanSoccerAnalysis$get_stadia(leagues, ids, names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{leagues}}{Leagues on which to filter. Accepts a character vector of length >= 1.}

\item{\code{ids}}{Stadium IDs on which to filter. Accepts a character vector of length >= 1.}

\item{\code{names}}{Stadium names on which to filter. Partial matches are accepted. Accepts a character vector of length >= 1.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_managers"></a>}}
\if{latex}{\out{\hypertarget{method-get_managers}{}}}
\subsection{Method \code{get_managers()}}{
Retrieves a data frame containing manager names and IDs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AmericanSoccerAnalysis$get_managers(leagues, ids, names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{leagues}}{Leagues on which to filter. Accepts a character vector of length >= 1.}

\item{\code{ids}}{Manager IDs on which to filter. Accepts a character vector of length >= 1.}

\item{\code{names}}{Manager names on which to filter. Partial matches are accepted. Accepts a character vector of length >= 1.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_referees"></a>}}
\if{latex}{\out{\hypertarget{method-get_referees}{}}}
\subsection{Method \code{get_referees()}}{
Retrieves a data frame containing referee names and IDs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AmericanSoccerAnalysis$get_referees(leagues, ids, names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{leagues}}{Leagues on which to filter. Accepts a character vector of length >= 1.}

\item{\code{ids}}{Referee IDs on which to filter. Accepts a character vector of length >= 1.}

\item{\code{names}}{Referee names on which to filter. Partial matches are accepted. Accepts a character vector of length >= 1.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_games"></a>}}
\if{latex}{\out{\hypertarget{method-get_games}{}}}
\subsection{Method \code{get_games()}}{
Retrieves a data frame containing game IDs, dates, opponents, scores, and other metadata.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AmericanSoccerAnalysis$get_games(
  leagues,
  game_ids,
  team_ids,
  team_names,
  seasons,
  stages
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{leagues}}{Leagues on which to filter. Accepts a character vector of length >= 1.}

\item{\code{game_ids}}{Game IDs on which to filter. Accepts a character vector of length >= 1.}

\item{\code{team_ids}}{Team IDs on which to filter. Accepts a character vector of length >= 1.}

\item{\code{team_names}}{Team names on which to filter. Partial matches and abbreviations are accepted. Accepts a character vector of length >= 1.}

\item{\code{seasons}}{Seasons on which to filter. Accepts a character or integer vector of length >= 1.}

\item{\code{stages}}{Stages (e.g., regular season, playoffs, etc.) on which to filter. See the \url{https://app.americansocceranalysis.com/api/v1/__swagger__/}{API documentation} for possible values. Accepts a character vector of length >= 1.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AmericanSoccerAnalysis$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
